jb: Make sure BytecodeBuilders are all counted
As part of the `connectTrees()` process whereby JitBuilder generated
IL is connected together into a full method's worth of IL, the initial
step is to count all the blocks that will be needed. With the newer jb2
implementation, it is possible that the BytecodeBuilders cannot be found
in this counting process because those BytecodeBuilders may be hidden
behind an IlBuilder that doesn't know how to "get to" the BytecodeBuilder.
So rather than rely on "finding" these builders, we can just proactively
add them all to the counting worklist, which is actually how the actual
"connecting" process works.

Signed-off-by: mstoodle <mstoodle@ca.ibm.com>