Change return type of TR_FrontEnd::getStringUTF8Length to int32_t
A String object that might consist of up to 2^31-1 chars could require
up to three times as many bytes as chars to be encoded in UTF-8.  The
current return type of getUTF8StringLength - intptr_t - is unable to
correctly represent all valid lengths in 32-bit mode.

This change makes the return type of the existing
TR_FrontEnd::getStringUTF8Length method int32_t.  It must only be used
in situations where it is known that the UTF-8 encoded length of the
String will not exceed 2^31-1 bytes.

This change also introduces a new getUTF8StringUnabbreviatedLength
method whose return type of uint64_t is able to represent all possible
UTF-8 encoded string lengths.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>