Propagate doNotProfile state when cloning nodes
BlockSplitter uses BlockCloner to split blocks along hot paths,
with the duplicated blocks forming the new hot path (that can then
be straightened) and the original blocks being pushed out to cold paths.

BlockCloner clones nodes using one of TR::Node's copy constructors,
which sets the doNotProfile on the clone's bytecode info. The copy
constructors are used all over the place, not just to create copies of
the original node but to also create new nodes. Setting the doNotProfile
flag avoids inadvertantly proflining semantically unrelated nodes that
happen to share the same bytecode, thereby polluting the profiling data.

Because of the above, when BlockSplitter runs, nothing on the hot
path will be profiled, and in cases where the cold paths are never
executed, we will completely miss out on profiling information. Because
BlockSplitter and BlockCloner are duplicating nodes they don't need to
worry about polluting the profiling data and should propagate the
doNotProfile flag state to the nodes they duplicate.

Signed-off-by: Younes Manton <ymanton@ca.ibm.com>