Remove isdigit macro redefs, replace with OMR_ISDIGIT
Changes here remove the redefinition of isdigit() in
the a2e ctype.h headers. In order to fully make this
work functionally, the references/usages of isdigit()
are replaced with OMR_ISDIGIT to properly recognize
ascii digit literals. This is done by conditionally
using __isdigit_a() on z/OS platforms.

Signed-off-by: Gaurav Chaudhari <gaurav.chaudhari@ibm.com>