Ensure simplification of compare op is valid
When the compare op is an unsigned compare, unless the select op's
DataType is TR::Address, simplifyISelectCompare() will recreate the
compare op with a BadILOp because compareOpCode() will not find a
valid comparison op of type TR_cmpNE. This will cause problems in
the comparison op simplifier afterward. Guards against this
simplification in these cases.