Improve x86 arrayset
For cases where the length of an arrayset is not known at compile-time,
or known to be >= 256 bytes, we generate a REP STOS sequence. This
is sub-optimal for small lengths as REP STOS has a relatively high
setup cost.

This patch implements a different sequence for a subset of the cases
that are currently handled by REP STOS. Specifically, if the arrayset
element size is 1 byte, it generates sequences to handle
the following lengths:

0 bytes
1-3 bytes, 2-3 stores, 1 branch
4-15 bytes, 4 stores, branch-free
16-31 bytes, 2 unaligned 16-byte stores, branch-free
32-63 bytes, 4 unaligned 16-byte stores, branch-free
>=64 bytes,
  1 unaligned 16-byte store,
  4 aligned 16-byte stores in a loop,
  3 aligned 16-byte stores in the residue
  1 unaligned 16-byte store in the residue

If the length is known to be >=64 only the loop will be generated.

Signed-off-by: Younes Manton <ymanton@ca.ibm.com>