x86: Insert padding after end-of-method call
In some scenarios, the last instruction in a method
can be a call instruction, such as when throwing
an exception. In such cases, the return address
will be beyond the end of the method body. This
commit adds padding to ensure the stack walker
does not throw an assertion due to an illegal
return address.

Issue: eclipse-openj9/openj9#21905

Signed-off-by: Bradley Wood <bradley.wood@ibm.com>