Set decimal precision when creating store for pdload
Partial redundancy optimizer add a store node for a load in some cases.
If the load is pdload, it creates a pdstore but never set the precision.
Later it call correctDecimalPrecision() which add a pdModifyPrecision
node to truncate the precision to zero.
This change fix the store node precision if required.

signed-off-by: Ehsan Kaini Far <ehsan.kianifar@gmail.com>