Use 64-bit instructions for length child and result regs
Last byte index for the arraycmplen as well as result register of the
node is 64-bit value. On 31-bit platform, it was generating
32-bit instructions for manipulating the index reg and result register
which could cause an issue where result register could contain a
garbage value in upper half. This commit fixes the instructions
generated for the arraycmplen SIMD helper.

Signed-off-by: Rahil Shah <rahil@ca.ibm.com>