RSInstructions encoding fix for IBM Z
This commit fixes the encoding issue reported in #5175 for RS
instructions on IBM Z. The second register of the register pair
was previously encoded irrespective of if it needed and it might
cause some unexpected behavior in future. This fix will skip
encoding the second register for the opcodes that do not use it.

Issue:#5175