RSInstructions encoding fix for IBM Z
This commit fixes the encoding issue reported in #5175 for RS
instructions on IBM Z. The odd register was previously encoded
irrespective of if it needed and it might cause some unexpected
behavior in future. This fix will skip encoding second/odd
registers for the opcodes that do not need it.

Issue:#5175