Save comp phase in a local rather than a stack allocated object
In `OMR::Optimizer::optimize()` we used to save the "compilation phase"
into a stack allocated object of type `TR::Compilation::CompilationPhaseScope`.
This happened in the constructor of the object. The destructor of the
object restores the previously saved state. This ensured that that if we
left the scope of the `OMR::Optimizer::optimize()` routine, the compilation
phase would be properly restored.
However, this is not what we want in case an exception is thrown from
within `OMR::Optimizer::optimize()` (or any method called from it). For
exceptions, we want to retain the compilation phase (namely the optimization)
that produced the exception, so that it can be properly reported at higher
levels, when the exception is caught.
This commit eliminates the `TR::Compilation::CompilationPhaseScope` object
and, instead, it stores the compilation phase into a local variable.
This local variable is used to restore the previous compilation phase when
we return from `OMR::Optimizer::optimize()`.
This commit also moves the `comp()->reportOptimizationPhase(optNum);`
statement from `performOptimization()` routine to an earlier location,
as soon as we determine that the current optimization is a proper optimization
and not an optimization group. This ensures that the "analysisPhase" reported
in case of a crash or exception is shown in the context of the current optimization
and not the previous one.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>