Disable matching of several patterns in CFG Simplification
Several patterns and corresponding transformations in CFG Simplification
have never been fully tested - in particular, simplifyNullToException,
simplifySimpleStore, simplifyCondStoreSequence,
simplifyInstanceOfTestToCheckcast and
simplifyBoundCheckWithThrowException.  This change disables those
transformations by default, allowing them to be enabled based on the
settings of various environment variables.  The intention is that they
should be more thoroughly tested and then reenabled as that testing and
any necessary bug fixes have been completed.

This change also fixes some typos in trace messages and changes the
declarations of some pointers that hold the results of calls to
feGetEnv from 'static char *' to 'static const char *' to emphasize
that the strings pointed to will not change.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>