Z: Add mask support in inlineVectorUnaryOp() and enable mask opcodes
Mask binary operations return the result of the operation in each lane
if the mask is true for that lane; otherwise, the result mirrors the
first child operand.
Add support for mask types in inlineVectorUnaryOp() function and
enables the following mask binary opcodes on the IBM Z platform: mor,
mand, mxor, mmin, and mmax.

signed-off-by: Ehsan Kiani Far <ehsan.kianifar@gmail.com>