RISC-V: specify CMAKE_C_COMPILER and CMAKE_CXX_COMPILER for CI builds
This commit explicitly specifies compilers to use for CI builds in order
to use ccache. The problem is that CMake looks for `cc` and `c++` when it
tries to autodetect the compiler. This should find `/usr/lib/ccache/cc`
as `/usr/lib/ccache` is prepended to the `PATH` but for some reason,
the RISC-V build agent does not have those and filesystem permissions
does not allow to create it from the job itself.

Therefore, we specify them explicitly.