Combine bu2i comparison optimization in z/codegen
The z/codegen skips evaluation of a comparison operation
if the first OR second child is a bu2i node and its
grandchild is a bloadi. This condition needs to be
modified from OR to AND (i.e so that both children of the
comparison operation satisfy all conditions). This fixes a
bug where the second comparison child has an
incompatible type with the bu2i node.

Fixes OpenJ9 issue:
https://github.com/eclipse-openj9/openj9/issues/22835

Signed-off-by: Dhruv Chopra <Dhruv.C.Chopra@ibm.com>