Fix reference counts on memrefs on Power
isLoadConstAndShift missed checking the reference count on a node. This
causes reference counts on the children of a node to be decremented
without evaluating the parent node. The parent node would then been
evaluated later on causing the children to be decremented again when
they should not. A check is added to ensure the parent node has a
reference count of 1 and that the parent has not already been evaluated
before proceeding with the optimization that skips evaluating the parent
node.

While fixing the above, it was also discovered that reference counts
were not properly decremented in two cases. There is an expectation
that the subTree node passed into the call to populateMemoryReference
gets decremented or gets recorded as _baseNode or _indexNode to be
decremented later. This was not happening in two places and has now been
fixed.

Signed-off-by: jimmyk <jimmyk@ca.ibm.com>